Working ADC led-dimmer and doubounced button

+++ adc.h
@@ -2,6 +2,6 @@
 #define _ADC_H_
 
 void adc_init(void);
-uint16_t get_ADC_val(void);
+uint8_t get_ADC_val(void);
 
 #endif //_ADC_H_
